pyproject.toml
[tool.poetry]
name = "last-project"
version = "0.1.0"
description = ""
authors = ["AdamTF21 <beka.0817a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
django = "^5.2"
telebot = "^0.0.5"
python-telegram-bot = "^22.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
drf-yasg = "^1.21.10"
aiogram = "^3.19.0"
python-dotenv = "^1.1.0"
pyproject-toml = "^0.1.0"
apscheduler = "^3.11.0"
asgiref = "^3.8.1"
django-jazzmin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
